requirements.txt
# Основные зависимости
aiogram>=3.0.0
aiohttp>=3.8.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-json-logger>=2.0.0
python-dotenv>=1.0.0

# База данных
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0

# Telegram
python-telegram-bot>=20.0
telethon>=1.28.0

# Обработка естественного языка
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.21.0
bitsandbytes>=0.39.0
peft>=0.4.0
sentencepiece>=0.1.99

# Утилиты
uvloop>=0.16.0
websockets>=11.0.3

# Для анализа данных и дашборда
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.13.0
dash>=2.9.0
dash-bootstrap-components>=1.3.0
openpyxl>=3.0.10
xlrd>=2.0.1

# Тестирование
pytest>=7.0.0
pytest-asyncio>=0.21.0
